backend-db-registration/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
minio==7.2.15
python-dotenv==1.0.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
boto3==1.34.34
pyyaml==6.0.2
pytest==8.0.2
pydantic==2.5.0
requests==2.31.0
